backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
langchain
langchain-community
sentence-transformers
torch
transformers
accelerate
llama-cpp-python
huggingface_hub
